# catalogue-backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Auto Parts Ordering API
  version: 1.0.0
  description: Backend contract for customer ordering, warehouse operations, and administration.
servers:
  - url: http://localhost:3000
...
paths:
...
  /api/orders:
    get:
      tags: [Orders, Admin, Warehouse]
      summary: Search orders for administration or fulfillment
      security: [{ adminKey: [] }, { warehouseKey: [] }]
      parameters:
        - { name: startDate, in: query, schema: { type: string, format: date } }
        - { name: endDate, in: query, schema: { type: string, format: date } }
        - { name: status, in: query, schema: { $ref: '#/components/schemas/OrderStatus' } }
        - { name: minPrice, in: query, schema: { type: number, minimum: 0 } }
        - { name: maxPrice, in: query, schema: { type: number, minimum: 0 } }
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/Offset' }
      responses:
        '200': { description: Matching order page }
        '401': { $ref: '#/components/responses/Error' }
    post:
      tags: [Orders]
      summary: Create, price, reserve, and authorize an order
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          schema: { type: string, maxLength: 128 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateOrder' }
      responses:
        '201':
          description: Newly authorized order
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DataOrder' }
        '200': { description: Idempotent replay }
        '402': { description: Payment declined; inventory was released, includes orderId }
        '409': { description: Insufficient inventory or reused idempotency key }
  /api/orders/{id}:
    get:
      tags: [Orders]
      summary: Get customer order status
      parameters:
        - { $ref: '#/components/parameters/OrderId' }
        - name: email
          in: query
          required: false
          schema: { type: string, format: email }
      responses:
        '200':
          description: Order
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DataOrder' }
        '401': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
    patch:
      tags: [Warehouse]
      summary: Ship an authorized order
      security: [{ warehouseKey: [] }]
      parameters: [{ $ref: '#/components/parameters/OrderId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [carrier, trackingNumber]
              properties:
                carrier: { type: string }
                trackingNumber: { type: string }
      responses:
        '200': { description: Shipped order }
        '401': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
  /api/orders/{id}/payment:
    post:
      tags: [Orders]
      summary: Retry payment for a PAYMENT_FAILED order
      parameters:
        - { $ref: '#/components/parameters/OrderId' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [customerEmail, payment]
              properties:
                customerEmail: { type: string, format: email }
                payment: { $ref: '#/components/schemas/Payment' }
      responses:
        '200':
          description: Authorized order
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DataOrder' }
        '402': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
...
  /api/warehouse/orders/{id}:
    get:
      tags: [Warehouse]
      summary: Get complete order
      security: [{ warehouseKey: [] }]
      parameters:
        - { $ref: '#/components/parameters/OrderId' }
      responses:
        '200':
          description: Order
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DataOrder' }
...
  /api/admin/orders:
    get:
      tags: [Admin]
      summary: Search orders
      security: [{ adminKey: [] }]
      parameters:
        - { name: dateFrom, in: query, schema: { type: string, format: date } }
        - { name: dateTo, in: query, schema: { type: string, format: date } }
        - { name: status, in: query, schema: { $ref: '#/components/schemas/OrderStatus' } }
        - { name: minPrice, in: query, schema: { type: number, minimum: 0 } }
        - { name: maxPrice, in: query, schema: { type: number, minimum: 0 } }
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/Offset' }
      responses:
        '200': { description: Matching order page }
  /api/admin/orders/{id}:
    get:
      tags: [Admin]
      summary: Get complete order details and history
      security: [{ adminKey: [] }]
      parameters: [{ $ref: '#/components/parameters/OrderId' }]
      responses:
        '200':
          description: Order
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DataOrder' }
components:
  securitySchemes:
    warehouseKey: { type: apiKey, in: header, name: X-API-Key }
    adminKey: { type: apiKey, in: header, name: X-API-Key }
  parameters:
...
    OrderId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
    Offset:
      name: offset
      in: query
      schema: { type: integer, minimum: 0, default: 0 }
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    OrderStatus:
      type: string
      enum: [PENDING_PAYMENT, AUTHORIZED, PAYMENT_FAILED, SHIPPED]
...
    Customer:
      type: object
      required: [name, email, address]
      properties:
        name: { type: string }
        email: { type: string, format: email }
        address:
          type: object
          required: [line1, city, state, postalCode, country]
          properties:
            line1: { type: string }
            line2: { type: string }
            city: { type: string }
            state: { type: string }
            postalCode: { type: string }
            country: { type: string, default: US }
    Payment:
      type: object
      required: [cardNumber, expirationDate, cardholderName]
      properties:
        cardNumber: { type: string, description: Luhn-valid; never persisted }
        expirationDate: { type: string, pattern: '^(0[1-9]|1[0-2])/\\d{4}$' }
        cardholderName: { type: string }
    CreateOrder:
      type: object
      required: [customer, items, payment]
      properties:
        customer: { $ref: '#/components/schemas/Customer' }
        items:
          type: array
          minItems: 1
          items:
            type: object
            required: [partNumber, quantity]
            properties:
              partNumber: { type: integer, minimum: 1 }
              quantity: { type: integer, minimum: 1, maximum: 1000 }
        payment: { $ref: '#/components/schemas/Payment' }
    OrderItem:
      type: object
      properties:
        partNumber: { type: integer }
        description: { type: string }
        unitPrice: { type: number }
        unitWeight: { type: number }
        quantity: { type: integer }
        lineTotal: { type: number }
    Order:
      type: object
      properties:
        id: { type: string, format: uuid }
        orderNumber: { type: integer }
        status: { $ref: '#/components/schemas/OrderStatus' }
        customer: { $ref: '#/components/schemas/Customer' }
        items: { type: array, items: { $ref: '#/components/schemas/OrderItem' } }
        subtotal: { type: number }
        shippingCharge: { type: number }
        totalAmount: { type: number }
        totalWeight: { type: number }
        payment:
          type: object
          properties:
            transactionId: { type: string }
            authorizationNumber: { type: [string, 'null'] }
            cardLast4: { type: [string, 'null'] }
        shipping:
          type: object
          properties:
            carrier: { type: [string, 'null'] }
            trackingNumber: { type: [string, 'null'] }
        createdAt: { type: string, format: date-time }
        authorizedAt: { type: [string, 'null'], format: date-time }
        shippedAt: { type: [string, 'null'], format: date-time }
...
    DataOrder:
      type: object
      properties:
        data: { $ref: '#/components/schemas/Order' }
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
            requestId: { type: string, format: uuid }
            details: {}